Fix umlauts not being set correctly
This commit fixes the bug that umlauts are not set correctly when settings strings from the settings.yaml file.
The bug occurs due to the default input encoding of latex.

After you have set the inputencoding to utf8 (l. 73) you can use umlatus in settings.yaml/in the source code.